huge update: date updating; safer editing; ...
When files are to be edited, they are copied to a temporary file for editing.
After editing the two files are diffed: if they are different, the temporary
file is moved to the entry file and the user is asked whether they want to
update the timestamp (the 'date' field).

The interactive number chooser now starts from 1 instead of 0. Return values are
the same.

Added a Y/N prompter function. Needs a docstring but it's easy enough to use
atm.

Added functions to retrieve front matter, file contents.

Also some cleanup: added self.message() method, error out on invalid command,
and more.